requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
deprecation==2.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.6
packaging==24.0
postgrest-py==0.10.6
psycopg2-binary==2.9.9
pydantic==1.10.15
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.1
StrEnum==0.4.15
typing_extensions==4.11.0
urllib3==2.2.1
